+++ src/main/java/nextstep/domain/theme/exception/ReservedThemeModifyException.java
@@ -1,4 +1,4 @@
-package nextstep.global.exceptions.exception;
+package nextstep.domain.theme.exception;
 
 public class ReservedThemeModifyException extends RuntimeException {
 }
