+++ src/main/java/nextstep/domain/theme/service/ThemeService.java
@@ -5,9 +5,9 @@ import nextstep.domain.reservation.repository.ReservationRepository;
 import nextstep.domain.theme.domain.Theme;
 import nextstep.domain.theme.dto.ThemeRequestDto;
 import nextstep.domain.theme.repository.ThemeRepository;
-import nextstep.global.exceptions.exception.DuplicatedNameThemeException;
-import nextstep.global.exceptions.exception.ReservedThemeModifyException;
-import nextstep.global.exceptions.exception.ThemeNotFoundException;
+import nextstep.domain.theme.exception.DuplicatedNameThemeException;
+import nextstep.domain.theme.exception.ReservedThemeModifyException;
+import nextstep.domain.theme.exception.ThemeNotFoundException;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
